move logic of translate pipe to base-translate pipe
preparation to add namespace translate pipe
